image/test/mochitest/test_smilEventsAllowedInImages.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=704482
-->
<head>
  <title>Test for Bug 704482</title>
  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/WindowSnapshot.js"></script>
  <script type="application/javascript" src="imgutils.js"></script>
  <script type="application/javascript" src="animationPolling.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=704482">
Mozilla Bug 704482
</a>
<p id="display"></p>

<div id="content">
  <div id="referenceDiv" style="height: 100px; width: 100px;
                                display: none; background: lime;">
  </div>
  <div id="animatedImage">
    <img id="img" style="display: none">
  </div>
  <div id="debug" style="display: none">
  </div>
</div>
<pre id="test">
<script type="text/javascript;version=1.8">
const FAILURE_TIMEOUT = 120000; // Fail early after 120 seconds (2 minutes)

var gIntervalId;
var gImageList = [
  "smil-event-begin.svg",
  "smil-event-beginEvent.svg",
  "smil-event-end.svg",
  "smil-event-endEvent.svg",
  "smil-event-repeat.svg",
  "smil-event-repeatEvent.svg",
  "smil-event-repeatN.svg",
  "smil-event-SVGLoad.svg",
];
var gNextImageIdx = 0;

const gImageElem = document.getElementById("img");
const gRefElem   = document.getElementById("referenceDiv");
const gDebugElem = document.getElementById("debug");

function preTestCleanup() {
  gImageElem.removeAttribute("src");
  gImageElem.style.display = "none";
  gRefElem.style.display   = "none";
  gDebugElem.style.display = "none";
}

function beginNextTest() {
  preTestCleanup();

  let curImageIdx = gNextImageIdx++;

  ok(curImageIdx < gImageList.length,
     "test sanity-check -- shouldn't iterate past end of JS array");

  gImageElem.setAttribute("src", gImageList[curImageIdx]);

  let nextAction = (gNextImageIdx == gImageList.length ? null : beginNextTest);

  let newTest = new AnimationTest(100, FAILURE_TIMEOUT, "referenceDiv",
                                  "img", "debug" , "", "", false,
                                  nextAction);
  newTest.beginTest();
}

function main()
{
  beginNextTest();
}

window.onload = main;
</script>
</pre>
</body>
</html>
